rename: 테스트 도메인 픽스쳐 클래스명 일괄 변경
- 기존 제공된 도메인 xxxTest 클래스를 픽스쳐로 사용함에 따라 이름 변경
