+++ src/test/java/qna/domain/AnswerFixture.java
@@ -1,0 +1,6 @@
+package qna.domain;
+
+public class AnswerFixture {
+    public static final Answer A1 = new Answer(UserFixture.JAVAJIGI, QuestionFixture.Q1, "Answers Contents1");
+    public static final Answer A2 = new Answer(UserFixture.SANJIGI, QuestionFixture.Q1, "Answers Contents2");
+}
