+++ src/test/java/qna/domain/QuestionFixture.java
@@ -1,0 +1,6 @@
+package qna.domain;
+
+public class QuestionFixture {
+    public static final Question Q1 = new Question("title1", "contents1").writeBy(UserFixture.JAVAJIGI);
+    public static final Question Q2 = new Question("title2", "contents2").writeBy(UserFixture.SANJIGI);
+}
